orchestrator/requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250909
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
astor==0.8.1
attrs==25.4.0
bitsandbytes==0.48.2
blake3==1.0.8
boto3==1.40.4
botocore==1.40.63
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.1
colorful==0.5.8
compressed-tensors==0.9.2
contourpy==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==3.6.0
deepspeed==0.18.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
evaluate==0.4.6
fastapi==0.120.3
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
filelock==3.20.0
flash_attn==2.7.4.post1
flashinfer-python==0.2.5+cu124torch2.6
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.3.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.42.1
google-genai==1.47.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mistral_common==1.8.5
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multiprocess==0.70.16
nanobind==2.9.2
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numba==0.61.0
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.6.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-exporter-prometheus==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post6
peft==0.17.1
pillow==12.0.0
platformdirs==4.5.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.2
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.51.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.2
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.2
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.4
torch==2.9.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.52.4
triton==3.5.0
trl==0.19.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
vllm==0.8.3
wandb==0.22.3
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
word2number==1.1
wrapt==2.0.0
xformers==0.0.29.post2
xgrammar==0.1.17
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0